refactor(analysis): split useLanguageAdapter into 3 focused modules (v3.3.3)
- languageAdapterTypes.ts: AdaptationStepId, AdaptationStep, AdaptationProgress, AdaptationResult
- languageAdapterPipeline.ts: reverseTranslate, reviewFidelity (stateless async helpers)
- useLanguageAdapter.ts: thin orchestration hook, imports types + pipeline, preserves public API
